// src-webviews/src/types/StaffJob.ts
export default interface StaffJob {
  id: string;
  name: string;
  posX: number;
  posY: number;
  posZ: number;
  salary: number;
  blipType: number;
  blipColor: number;
  blipName: string;
  vehicleRentModel: string;
  vehicleRentValue: number;
  vehicleRentPosX: number;
  vehicleRentPosY: number;
  vehicleRentPosZ: number;
  vehicleRentRotR: number;
  vehicleRentRotP: number;
  vehicleRentRotY: number;
};